refactor: extract Console video controller
