+++ backlog/tasks/task-3070 - chat_screen-size-ratchet-red-on-dev-after-console-decomposition-wave-3.md
@@ -18,12 +18,12 @@ priority: medium
 
 <!-- SECTION:DESCRIPTION:BEGIN -->
 The Console's one-way architecture ratchets are materially red on verified current
-`origin/dev`: `tldw_chatbook/UI/Screens/chat_screen.py` is 22,338 lines against a
-17,727-line ceiling, and `ChatScreen` has 705 direct methods against a 593-method
-ceiling. This task was originally filed at a 21-line overage after wave 3; later waves
+`origin/dev` after TASK-3070.2: `tldw_chatbook/UI/Screens/chat_screen.py` is 22,172
+lines against a 17,727-line ceiling, and `ChatScreen` has 712 direct methods against a
+593-method ceiling. This task was originally filed at a 21-line overage after wave 3; later waves
 lowered the allowed ceiling while unrelated feature work kept accumulating in the
-screen. The result is now 4,611 lines and 112 methods over budget, not transitional
-noise. Resolve it through the reviewed Wave 6 controller extractions and lower the
+screen. The result is now 4,445 lines and 119 methods over budget, not transitional
+noise. Resolve it through the amended Wave 6 controller extractions and lower the
 ratchets to the exact earned counts; never raise either budget to accept the growth.
 <!-- SECTION:DESCRIPTION:END -->
 
@@ -31,9 +31,9 @@ ratchets to the exact earned counts; never raise either budget to accept the gro
 <!-- AC:BEGIN -->
 - [ ] #1 On the final rebased dev base, both the `chat_screen.py` line-count and `ChatScreen` method-count ratchets pass without increasing either budget.
 - [ ] #2 The source-inspected Wave 6 move/delegate/stay inventory is enforced by AST ownership tests, and every retained screen delegate is framework-required, has a real caller, and its complete definition span (decorators excluded) is bounded to five physical source lines.
-- [ ] #3 Image/H3, video, conversation-browser, retrieval/RAG, skill, and character ownership moves to the reviewed controllers without changing DOM ownership, persistence ordering, worker groups, cancellation identity, remount/shutdown behavior, privacy, or user-visible outcomes.
+- [ ] #3 Image/H3, video, conversation-browser, retrieval/RAG, skill, character, fleet/wake, first-chat, and auto-speak ownership moves to the reviewed controllers without changing DOM ownership, persistence ordering, worker groups, cancellation identity, remount/shutdown behavior, privacy, or user-visible outcomes.
 - [ ] #4 Every extracted family has isolated controller coverage using plain fakes without mounting Textual, including the Workspace browser extension, and the mounted product suites continue to cover screen/region integration.
-- [ ] #5 The coordinated child tasks TASK-3070.1 through TASK-3070.8 are completed independently with their focused verification evidence before this parent is closed.
+- [ ] #5 The coordinated child tasks TASK-3070.1 through TASK-3070.11 are completed independently with their focused verification evidence before this parent is closed.
 <!-- AC:END -->
 
 ## Implementation Plan
@@ -44,10 +44,10 @@ ADR path: N/A (governed by `Docs/superpowers/specs/2026-08-02-screen-decompositi
 Reason: This is the next implementation wave of the existing, approved Console decomposition architecture; it does not introduce a new runtime, storage, security, or cross-module policy.
 
 1. Complete TASK-3070.1: record the rebased line/method baseline and lock the reviewed source-inspected ownership inventory, residue budget, and AST rules before production movement.
-2. Complete TASK-3070.2 through TASK-3070.7 one PR at a time after its predecessor merges: characterize the real product boundary, add isolated no-mount controller tests, extract one family with named late-bound dependencies, and prove its screen delegates and invariants.
+2. Complete TASK-3070.2 through TASK-3070.10 one PR at a time after its predecessor merges: characterize the real product boundary, add isolated no-mount controller tests, extract one family with named late-bound dependencies, and prove its screen delegates and invariants.
 3. Extend the existing Workspace controller as the single conversation-browser state owner while preserving collapse/config and activation/resume behavior; keep Textual decorators, direct DOM, modal/picker presentation, and OS-player launch on the screen or existing region widgets.
 4. After each extraction, run its isolated controller tests, focused product/mounted suites, AST ownership/dependency checks, required mutations, static checks, and privacy/diff gates before beginning the next child.
-5. Complete TASK-3070.8: rebase onto final `origin/dev`, repeat the baseline comparison, lower both ratchets to the exact earned counts, update canonical decomposition progress and every task's notes, run the full repository DoD, and only then close this parent.
+5. Complete TASK-3070.11: rebase onto final `origin/dev`, repeat the baseline comparison, lower both ratchets to the exact earned counts, update canonical decomposition progress and every task's notes, run the final repository DoD, and only then close this parent.
 <!-- SECTION:PLAN:END -->
 
 ## Implementation Notes
