+++ backlog/tasks/task-3070.10 - Extract-Console-auto-speak-ownership.md
@@ -1,0 +1,30 @@
+---
+id: TASK-3070.10
+title: Extract Console auto-speak ownership
+status: To Do
+assignee:
+  - '@codex'
+created_date: '2026-08-15 00:15'
+labels:
+  - console
+  - speech
+dependencies:
+  - TASK-3070.9
+references:
+  - Docs/superpowers/specs/2026-08-13-console-decomposition-wave6-design.md
+parent_task_id: TASK-3070
+priority: high
+---
+
+## Description
+
+<!-- SECTION:DESCRIPTION:BEGIN -->
+Move post-baseline auto-speak destination and control-state ownership into the existing HandsFree controller while retaining Textual event entry points on ChatScreen.
+<!-- SECTION:DESCRIPTION:END -->
+
+## Acceptance Criteria
+<!-- AC:BEGIN -->
+- [ ] #1 Auto-speak policy is HandsFree-owned and the three decorated screen entry points are bounded delegates
+- [ ] #2 Destination, resume, retry, queue, and presentation behavior remain unchanged
+- [ ] #3 Isolated no-mount policy tests and focused speech integration tests pass
+<!-- AC:END -->
