+++ backlog/tasks/task-3070.11 - Close-Console-decomposition-Wave-6-ratchet.md
@@ -1,11 +1,10 @@
 ---
-id: TASK-3070.8
+id: TASK-3070.11
 title: Close Console decomposition Wave 6 ratchet
 status: To Do
 assignee:
   - '@codex'
-created_date: '2026-08-13 17:09'
-updated_date: '2026-08-13 17:16'
+created_date: '2026-08-15 00:15'
 labels:
   - console
   - architecture-gate
@@ -17,6 +16,9 @@ dependencies:
   - TASK-3070.5
   - TASK-3070.6
   - TASK-3070.7
+  - TASK-3070.8
+  - TASK-3070.9
+  - TASK-3070.10
 references:
   - Docs/superpowers/specs/2026-08-13-console-decomposition-wave6-design.md
 parent_task_id: TASK-3070
@@ -32,6 +34,6 @@ Rebase the completed Wave 6 extractions, lower the one-way ratchet to the exact 
 ## Acceptance Criteria
 <!-- AC:BEGIN -->
 - [ ] #1 Final rebased ChatScreen line and method counts pass ratchets without increasing a budget
-- [ ] #2 All focused, mounted, architecture, static, privacy, and full-suite gates pass
+- [ ] #2 All approved focused, mounted, architecture, static, privacy, and final integration gates pass
 - [ ] #3 TASK-3070 and all completed child tasks contain implementation notes and satisfy repository DoD
 <!-- AC:END -->
