+++ backlog/tasks/task-3070.3 - Extract-Console-video-controller.md
@@ -1,11 +1,11 @@
 ---
 id: TASK-3070.3
 title: Extract Console video controller
-status: To Do
+status: Done
 assignee:
   - '@codex'
 created_date: '2026-08-13 17:08'
-updated_date: '2026-08-13 17:14'
+updated_date: '2026-08-15 00:48'
 labels:
   - console
   - video
@@ -25,7 +25,46 @@ Move generated-video lifecycle, publication, persistence, and orchestration owne
 
 ## Acceptance Criteria
 <!-- AC:BEGIN -->
-- [ ] #1 Video move inventory is controller-owned and screen entry points are bounded delegates
-- [ ] #2 Cancellation identity, worker groups, storage/publication order, remount drains, and presentation behavior are unchanged
-- [ ] #3 Isolated controller tests run without mounting Textual
+- [x] #1 Video move inventory is controller-owned and screen entry points are bounded delegates
+- [x] #2 Cancellation identity, worker groups, storage/publication order, remount drains, and presentation behavior are unchanged
+- [x] #3 Isolated controller tests run without mounting Textual
 <!-- AC:END -->
+
+## Implementation Plan
+
+<!-- SECTION:PLAN:BEGIN -->
+Detailed executable plan:
+`Docs/superpowers/plans/2026-08-14-task-3070-3-console-video-controller.md`
+
+1. Lock the approved post-image baseline amendment and prove the expanded inventory
+   still clears both ratchets before changing video production code.
+2. Add failing no-mount controller, compatibility, direct-wiring, cancellation,
+   persistence-order, and drain contracts.
+3. Move exactly the reviewed 31 non-DOM video methods and eight state defaults
+   into `ConsoleVideoController`; keep the two registry-bound screen delegates
+   and both presentation seams.
+4. Wire the controller in `wiring.py`, route MessageController
+   play/save/regenerate directly to it, and preserve all screen compatibility
+   names through read/write descriptors.
+5. Retarget direct private tests without changing assertions, mutation-check the
+   lifecycle invariants, and run only the focused modified-functionality suites.
+6. Complete static/privacy/diff verification, independent review, and task
+   closeout.
+
+ADR required: no
+
+ADR path: N/A
+
+Reason: this implements the approved Wave 6 ownership boundary and introduces no
+new storage, security, provider, runtime, or cross-module policy.
+<!-- SECTION:PLAN:END -->
+
+## Implementation Notes
+
+<!-- SECTION:NOTES:BEGIN -->
+Extracted the reviewed generated-video lifecycle into a no-DOM ConsoleVideoController with eight read/write compatibility descriptors and two bounded registry delegates. Wired MessageController play/save/regenerate directly to the controller, preserved the durable storage identity and publication/cancellation order, and retargeted private tests to the defining owner without changing behavior assertions.
+
+Verification: 335 focused tests passed, 2 Windows-only skips, and 1 pre-existing dependency warning; four mutation probes failed as intended and were restored; two persistent-diagnostic governance nodes passed after moving the metadata-only diagnostic owner; Ruff check passed for all extracted/retargeted files and chat_screen.py retained its exact 28-finding pre-task baseline; Bandit, isolated py_compile, generated diagnostic equality, diff/privacy/artifact checks passed. The repository-wide suite was intentionally not run per owner direction.
+
+ADR required: no. This implements the approved Wave 6 boundary without changing storage, provider, security, runtime, or cross-module policy.
+<!-- SECTION:NOTES:END -->
