+++ backlog/tasks/task-3070.4 - Consolidate-Console-conversation-browser-into-Workspace-controller.md
@@ -25,7 +25,7 @@ Move conversation row, cache, filtering, and search ownership out of ChatScreen 
 
 ## Acceptance Criteria
 <!-- AC:BEGIN -->
-- [ ] #1 ConsoleWorkspaceController owns one canonical browser query/timer/token/result lifecycle and the screen search handler is a bounded delegate
+- [ ] #1 ConsoleWorkspaceController owns one canonical browser query/timer/token/result lifecycle, including durable unseen-marker projection, and the screen search handler is a bounded delegate
 - [ ] #2 Rich browser rows are canonical; legacy Workspace rows are an explicit read/write projection, and no legacy refresh path remains as a second writer
 - [ ] #3 Existing and newer scalar browser compatibility names read and write the same Workspace-owned state while collapse/config and activation/resume behavior remain unchanged
 - [ ] #4 Existing Workspace controller tests cover the browser transitions and row projections without mounting Textual, and AST checks reject `query_one` in every moved browser method
