+++ backlog/tasks/task-3070.8 - Extract-Console-fleet-and-wake-lifecycle-controller.md
@@ -1,0 +1,31 @@
+---
+id: TASK-3070.8
+title: Extract Console fleet and wake lifecycle controller
+status: To Do
+assignee:
+  - '@codex'
+created_date: '2026-08-13 17:09'
+updated_date: '2026-08-15 00:15'
+labels:
+  - console
+  - architecture-gate
+dependencies:
+  - TASK-3070.7
+references:
+  - Docs/superpowers/specs/2026-08-13-console-decomposition-wave6-design.md
+parent_task_id: TASK-3070
+priority: high
+---
+
+## Description
+
+<!-- SECTION:DESCRIPTION:BEGIN -->
+Move post-baseline fleet completion, wake coordination, unseen-marker policy, teardown, and survivor-tick lifecycle out of ChatScreen without changing delivery, cancellation, repaint, or persistence behavior.
+<!-- SECTION:DESCRIPTION:END -->
+
+## Acceptance Criteria
+<!-- AC:BEGIN -->
+- [ ] #1 Fleet/wake move inventory is controller-owned with zero controller DOM access
+- [ ] #2 First-signal delivery, unseen-marker, teardown, and survivor-tick behavior remain unchanged
+- [ ] #3 Isolated no-mount lifecycle tests and focused integration tests pass
+<!-- AC:END -->
