+++ backlog/tasks/task-3070.9 - Extract-Console-first-chat-handoff-ownership.md
@@ -1,0 +1,30 @@
+---
+id: TASK-3070.9
+title: Extract Console first-chat handoff ownership
+status: To Do
+assignee:
+  - '@codex'
+created_date: '2026-08-15 00:15'
+labels:
+  - console
+  - handoff
+dependencies:
+  - TASK-3070.8
+references:
+  - Docs/superpowers/specs/2026-08-13-console-decomposition-wave6-design.md
+parent_task_id: TASK-3070
+priority: high
+---
+
+## Description
+
+<!-- SECTION:DESCRIPTION:BEGIN -->
+Move post-baseline first-chat claim fencing rollback acknowledgement and resynchronization policy out of ChatScreen while preserving current session and focus behavior.
+<!-- SECTION:DESCRIPTION:END -->
+
+## Acceptance Criteria
+<!-- AC:BEGIN -->
+- [ ] #1 First-chat handoff move inventory is controller-owned and screen callbacks remain presentation-only
+- [ ] #2 Claim fencing rollback acknowledgement retry privacy and focus restoration behavior remain unchanged
+- [ ] #3 Isolated no-mount handoff tests and focused integration tests pass
+<!-- AC:END -->
